+++ lib/bali.rb
@@ -15,6 +15,7 @@ require 'bali/path_helper'
 require 'bali/form_helper'
 require 'bali/auto_submit_select_helper'
 require 'bali/concerns/numeric_attributes_with_commas'
+require 'bali/concerns/date_range_attribute'
 
 builder_helpers = File.join(File.dirname(__FILE__), 'bali/form_builder', '*_fields.rb')