+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.61.5] - 2023-03-31
+
+### Added
+
+- `Bali::Concerns::DateRangeAttribute` concern. This concern allows to define date range attributes, for example, `date_range_attribute :date_range, default: Time.zone.now.all_day`
+- `max_date` option to `date_field_group` and `date_field`. you to set a maximum date that can be selected
+
 ## [0.61.4] - 2023-03-30
 
 ### Changed